Blessing Project Application
There are 2 Apps:
1. The Reader (Blessing Pro)
2. The Story Maker
(These apps are made with Python programming script)
The app has contain 7 stories. If you wanted to make new one, you can use Story Maker. However there are steps that you need to follow through, in order to make it working in the reader app. The steps will be documented and explain in How To Make a Story documentation.